presidio_guardrail/requirements.txt
nemoguardrails>=0.10.0
presidio-analyzer>=2.2
presidio-anonymizer>=2.2
spacy>=3.7
langchain-openai>=0.1.0
python-dotenv>=1.0.0
streamlit>=1.30.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp-proto-http>=1.20.0
# After install: python -m spacy download en_core_web_lg
